--- core/domain/src/main/java/kky/flab/lookaround/core/domain/MyClass.kt
@@ -1,4 +1,0 @@
-package kky.flab.lookaround.core.domain
-
-class MyClass {
-}
