+++ core/domain/src/main/java/kky/flab/lookaround/feature/domain/MyClass.kt
@@ -1,0 +1,4 @@
+package kky.flab.lookaround.feature.domain
+
+class MyClass {
+}
