+++ feature/main/src/main/java/kky/flab/lookaround/feature/main/MainActivity.kt
@@ -1,0 +1,11 @@
+package kky.flab.lookaround.feature.main
+
+import android.os.Bundle
+import androidx.appcompat.app.AppCompatActivity
+
+class MainActivity : AppCompatActivity() {
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        setContentView(R.layout.activity_main)
+    }
+}
