+++ feature/main/src/main/res/color/bottom_navigation_icon_color.xml
@@ -1,0 +1,5 @@
+<?xml version="1.0" encoding="utf-8"?>
+<selector xmlns:android="http://schemas.android.com/apk/res/android">
+    <item android:state_checked="true" android:color="@color/primaryColor"/>
+    <item android:state_checked="false" android:color="@color/gray"/>
+</selector>
