+++ feature/main/src/main/res/drawable/main_bottom_home.xml
@@ -1,0 +1,5 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android" android:height="24dp" android:viewportHeight="24" android:viewportWidth="24" android:width="24dp">
+      
+    <path android:fillColor="#FF000000" android:pathData="M23.121,9.069 L15.536,1.483a5.008,5.008 0,0 0,-7.072 0L0.879,9.069A2.978,2.978 0,0 0,0 11.19v9.817a3,3 0,0 0,3 3L21,24.007a3,3 0,0 0,3 -3L24,11.19A2.978,2.978 0,0 0,23.121 9.069ZM15,22.007L9,22.007L9,18.073a3,3 0,0 1,6 0ZM22,21.007a1,1 0,0 1,-1 1L17,22.007L17,18.073a5,5 0,0 0,-10 0v3.934L3,22.007a1,1 0,0 1,-1 -1L2,11.19a1.008,1.008 0,0 1,0.293 -0.707L9.878,2.9a3.008,3.008 0,0 1,4.244 0l7.585,7.586A1.008,1.008 0,0 1,22 11.19Z"/>
+    
+</vector>
