+++ feature/main/src/main/res/drawable/main_bottom_record.xml
@@ -1,0 +1,9 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="24dp"
+    android:height="24dp"
+    android:viewportWidth="24"
+    android:viewportHeight="24">
+  <path
+      android:fillColor="#FF000000"
+      android:pathData="m23.837,11.479l-1.852,-10.65c-0.095,-0.544 -0.611,-0.907 -1.157,-0.814 -0.544,0.095 -0.909,0.613 -0.814,1.157l1.852,10.65c0.01,0.059 0.014,0.119 0.023,0.178h-3.89c-1.103,0 -2,0.897 -2,2s-0.897,2 -2,2h-4c-1.103,0 -2,-0.897 -2,-2s-0.897,-2 -2,-2h-3.89c0.009,-0.059 0.013,-0.12 0.023,-0.179L3.985,1.171c0.095,-0.544 -0.27,-1.062 -0.814,-1.157 -0.548,-0.093 -1.062,0.271 -1.157,0.814L0.163,11.478c-0.082,0.469 -0.163,1.785 -0.163,1.885v5.637c0,2.757 2.243,5 5,5h14c2.757,0 5,-2.243 5,-5v-5.637c0,-0.1 -0.081,-1.417 -0.163,-1.885ZM19,22L5,22c-1.654,0 -3,-1.346 -3,-3v-5h4c0,2.206 1.794,4 4,4h4c2.206,0 4,-1.794 4,-4h4v5c0,1.654 -1.346,3 -3,3ZM6.22,0.818c0.088,-0.474 0.501,-0.818 0.983,-0.818h9.604c0.482,0 0.896,0.344 0.984,0.819 0.113,0.614 -0.359,1.181 -0.984,1.181L7.204,2c-0.625,0 -1.097,-0.567 -0.983,-1.182ZM4.74,8.818c0.088,-0.474 0.501,-0.818 0.983,-0.818h12.556c0.482,0 0.896,0.344 0.984,0.819 0.113,0.614 -0.359,1.181 -0.984,1.181L5.724,10c-0.625,0 -1.097,-0.567 -0.983,-1.182ZM5.48,4.818c0.088,-0.474 0.501,-0.818 0.983,-0.818h11.08c0.482,0 0.896,0.344 0.984,0.819 0.113,0.614 -0.359,1.181 -0.984,1.181L6.464,6c-0.625,0 -1.097,-0.567 -0.983,-1.182Z"/>
+</vector>
