+++ feature/main/src/main/res/drawable/main_bottom_setting.xml
@@ -1,0 +1,7 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android" android:height="24dp" android:viewportHeight="24" android:viewportWidth="24" android:width="24dp">
+      
+    <path android:fillColor="#FF000000" android:pathData="M12,8a4,4 0,1 0,4 4A4,4 0,0 0,12 8ZM12,14a2,2 0,1 1,2 -2A2,2 0,0 1,12 14Z"/>
+      
+    <path android:fillColor="#FF000000" android:pathData="M21.294,13.9l-0.444,-0.256a9.1,9.1 0,0 0,0 -3.29l0.444,-0.256a3,3 0,1 0,-3 -5.2l-0.445,0.257A8.977,8.977 0,0 0,15 3.513L15,3A3,3 0,0 0,9 3v0.513A8.977,8.977 0,0 0,6.152 5.159L5.705,4.9a3,3 0,0 0,-3 5.2l0.444,0.256a9.1,9.1 0,0 0,0 3.29l-0.444,0.256a3,3 0,1 0,3 5.2l0.445,-0.257A8.977,8.977 0,0 0,9 20.487L9,21a3,3 0,0 0,6 0v-0.513a8.977,8.977 0,0 0,2.848 -1.646l0.447,0.258a3,3 0,0 0,3 -5.2ZM18.746,10.124a7.048,7.048 0,0 1,0 3.75,1 1,0 0,0 0.464,1.133l1.084,0.626a1,1 0,0 1,-1 1.733l-1.086,-0.628a1,1 0,0 0,-1.215 0.165,6.984 6.984,0 0,1 -3.243,1.875 1,1 0,0 0,-0.751 0.969L12.999,21a1,1 0,0 1,-2 0L10.999,19.748a1,1 0,0 0,-0.751 -0.969A6.984,6.984 0,0 1,7.006 16.9a1,1 0,0 0,-1.215 -0.165l-1.084,0.627a1,1 0,1 1,-1 -1.732l1.084,-0.626a1,1 0,0 0,0.464 -1.133,7.048 7.048,0 0,1 0,-3.75A1,1 0,0 0,4.79 8.992L3.706,8.366a1,1 0,0 1,1 -1.733l1.086,0.628A1,1 0,0 0,7.006 7.1a6.984,6.984 0,0 1,3.243 -1.875A1,1 0,0 0,11 4.252L11,3a1,1 0,0 1,2 0L13,4.252a1,1 0,0 0,0.751 0.969A6.984,6.984 0,0 1,16.994 7.1a1,1 0,0 0,1.215 0.165l1.084,-0.627a1,1 0,1 1,1 1.732l-1.084,0.626A1,1 0,0 0,18.746 10.125Z"/>
+    
+</vector>
