+++ feature/main/src/main/res/menu/bottom_navigation_menu.xml
@@ -1,0 +1,10 @@
+<?xml version="1.0" encoding="utf-8"?>
+<menu xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto">
+    <item android:title="홈" android:icon="@drawable/main_bottom_home"
+        app:showAsAction="ifRoom"/>
+    <item android:title="기록" android:icon="@drawable/main_bottom_record"
+        app:showAsAction="ifRoom"/>
+    <item android:title="설정" android:icon="@drawable/main_bottom_setting"
+        app:showAsAction="ifRoom"/>
+</menu>
