--- feature/main/src/main/res/values/themes.xml
@@ -1,5 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<resources>
-
-    <style name="Theme.Lookaround" parent="Theme.AppCompat.Light.NoActionBar" />
-</resources>
